Galaxy.txt
Gives you the Galaxy chair, a limited item. It won't be available in your inventory but everywhere towards the server. The game also saves its existence meaning you gotta execute this script only once in order to have the chair. If you change it  you need to re execute it.